# docs/swagger-api.yml
openapi: 3.0.3
info:
  title: Architecture Hybride - API Gateway
  description: |
    Documentation complète de l'API Gateway pour l'architecture hybride LOG430.
    
    Cette API expose un routage intelligent entre :
    - Système de base (Legacy) pour les consoles POS et Maison Mère
    - Microservices modernes pour les API et interfaces Web/Mobile
    
    ## Routage Intelligent
    
    Le routage se fait automatiquement selon :
    - **Header `X-Client-Type`** : pos, maisonmere, api, web
    - **Path prefix** : /pos/, /maisonmere/, /api/v2/, /web/
    
    ## Authentification
    
    - **API Key** : Header `X-API-Key` pour les intégrations externes
    - **CORS** : Configuré pour les domaines autorisés
    - **Rate Limiting** : 200 req/min par défaut
    
  version: 2.0.0
  contact:
    name: LOG430 Lab TB
    url: http://localhost:9000
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8001
    description: Kong Gateway (Production)
  - url: http://localhost:9000
    description: Hybrid Router (Development)

tags:
  - name: Console POS
    description: Point de vente - utilise Legacy + Stock microservice
  - name: Console Maison Mère
    description: Administration - utilise Legacy + Reporting microservice
  - name: API Moderne
    description: Intégrations externes - utilise tous les microservices
  - name: Web/Mobile
    description: Interface utilisateur - utilise tous les microservices
  - name: Health & Monitoring
    description: Endpoints de santé et monitoring

paths:
  # =======================================
  # CONSOLE POS
  # =======================================
  /pos/produits:
    get:
      tags: [Console POS]
      summary: Liste des produits (Legacy)
      description: Récupère la liste des produits via le système legacy
      parameters:
        - name: X-Client-Type
          in: header
          required: true
          schema:
            type: string
            enum: [pos]
          example: pos
        - name: search
          in: query
          schema:
            type: string
          description: Recherche par nom de produit
        - name: category
          in: query
          schema:
            type: string
          description: Filtrer par catégorie
      responses:
        '200':
          description: Liste des produits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProduitLegacy'
        '429':
          description: Rate limit dépassé
        '500':
          description: Erreur serveur

  /pos/stock/{productId}:
    get:
      tags: [Console POS]
      summary: Stock temps réel (Microservice)
      description: Vérifie le stock en temps réel via le microservice Stock
      parameters:
        - name: X-Client-Type
          in: header
          required: true
          schema:
            type: string
            enum: [pos]
        - name: productId
          in: path
          required: true
          schema:
            type: integer
          description: ID du produit
      responses:
        '200':
          description: Niveau de stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'

  /pos/ventes:
    post:
      tags: [Console POS]
      summary: Créer une vente (Legacy)
      description: Enregistre une nouvelle vente via le système legacy
      parameters:
        - name: X-Client-Type
          in: header
          required: true
          schema:
            type: string
            enum: [pos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VenteRequest'
      responses:
        '201':
          description: Vente créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VenteResponse'

  # =======================================
  # CONSOLE MAISON MÈRE
  # =======================================
  /maisonmere/dashboard:
    get:
      tags: [Console Maison Mère]
      summary: Dashboard principal (Legacy)
      description: Données du dashboard via le système legacy
      parameters:
        - name: X-Client-Type
          in: header
          required: true
          schema:
            type: string
            enum: [maisonmere]
      responses:
        '200':
          description: Données du dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'

  /maisonmere/reports/analytics:
    get:
      tags: [Console Maison Mère]
      summary: Rapports avancés (Microservice)
      description: Analytics et BI via le microservice Reporting
      parameters:
        - name: X-Client-Type
          in: header
          required: true
          schema:
            type: string
            enum: [maisonmere]
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month, year]
          description: Période du rapport
      responses:
        '200':
          description: Rapport d'analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsReport'

  # =======================================
  # API MODERNE
  # =======================================
  /api/v2/produits:
    get:
      tags: [API Moderne]
      summary: Catalogue produits (Microservice)
      description: API REST moderne du catalogue produits
      parameters:
        - name: X-Client-Type
          in: header
          schema:
            type: string
            enum: [api]
        - name: X-API-Key
          in: header
          required: true
          schema:
            type: string
          description: Clé API pour authentification
      responses:
        '200':
          description: Catalogue produits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProduitModerne'
        '401':
          description: API Key invalide
        '429':
          description: Rate limit dépassé
    
    post:
      tags: [API Moderne]
      summary: Créer un produit (Microservice)
      description: Créer un nouveau produit via l'API moderne
      parameters:
        - name: X-API-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProduitCreateRequest'
      responses:
        '201':
          description: Produit créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProduitModerne'

  /api/v2/analytics:
    get:
      tags: [API Moderne]
      summary: Analytics API (Microservice)
      description: Données d'analytics pour intégrations externes
      parameters:
        - name: X-API-Key
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Données analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsAPI'

  # =======================================
  # WEB/MOBILE
  # =======================================
  /web/catalog:
    get:
      tags: [Web/Mobile]
      summary: Catalogue web (Microservice)
      description: Catalogue optimisé pour interface web/mobile
      parameters:
        - name: X-Client-Type
          in: header
          schema:
            type: string
            enum: [web]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Catalogue paginé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'

  # =======================================
  # HEALTH & MONITORING
  # =======================================
  /health:
    get:
      tags: [Health & Monitoring]
      summary: Health check global
      description: Vérifie la santé de tous les services
      responses:
        '200':
          description: Statut des services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheck'

  /metrics:
    get:
      tags: [Health & Monitoring]
      summary: Métriques Prometheus
      description: Métriques au format Prometheus
      responses:
        '200':
          description: Métriques
          content:
            text/plain:
              schema:
                type: string

components:
  schemas:
    ProduitLegacy:
      type: object
      properties:
        id:
          type: integer
          example: 123
        nom:
          type: string
          example: "iPhone 15"
        prix:
          type: number
          format: float
          example: 999.99
        categorie:
          type: string
          example: "Smartphones"
        description:
          type: string
          example: "Dernier iPhone avec puce A17"

    ProduitModerne:
      type: object
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: "iPhone 15"
        price:
          type: number
          format: float
          example: 999.99
        category:
          type: string
          example: "Smartphones"
        description:
          type: string
          example: "Latest iPhone with A17 chip"
        images:
          type: array
          items:
            type: string
          example: ["https://example.com/image1.jpg"]
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Stock:
      type: object
      properties:
        productId:
          type: integer
          example: 123
        available:
          type: integer
          example: 5
        reserved:
          type: integer
          example: 2
        total:
          type: integer
          example: 7
        lastUpdate:
          type: string
          format: date-time

    VenteRequest:
      type: object
      required:
        - productId
        - quantity
        - customerId
      properties:
        productId:
          type: integer
          example: 123
        quantity:
          type: integer
          example: 1
        customerId:
          type: integer
          example: 456
        paymentMethod:
          type: string
          enum: [cash, card, mobile]
          example: "card"

    VenteResponse:
      type: object
      properties:
        id:
          type: integer
          example: 789
        productId:
          type: integer
          example: 123
        quantity:
          type: integer
          example: 1
        total:
          type: number
          format: float
          example: 999.99
        status:
          type: string
          enum: [pending, completed, cancelled]
          example: "completed"
        timestamp:
          type: string
          format: date-time

    Dashboard:
      type: object
      properties:
        totalSales:
          type: number
          format: float
          example: 45000.50
        totalOrders:
          type: integer
          example: 256
        topProducts:
          type: array
          items:
            $ref: '#/components/schemas/ProduitLegacy'
        recentActivity:
          type: array
          items:
            type: object

    AnalyticsReport:
      type: object
      properties:
        period:
          type: string
          example: "month"
        metrics:
          type: object
          properties:
            revenue:
              type: number
              format: float
              example: 125000.75
            orders:
              type: integer
              example: 1250
            avgOrderValue:
              type: number
              format: float
              example: 100.00
        trends:
          type: array
          items:
            type: object
        predictions:
          type: object

    AnalyticsAPI:
      type: object
      properties:
        timeRange:
          type: string
          example: "last_30_days"
        data:
          type: object
          additionalProperties: true
        exportUrl:
          type: string
          example: "https://api.example.com/export/12345"

    CatalogResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProduitModerne'
        pagination:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            total:
              type: integer
              example: 500
            totalPages:
              type: integer
              example: 25

    ProduitCreateRequest:
      type: object
      required:
        - name
        - price
        - category
      properties:
        name:
          type: string
          example: "Nouveau Produit"
        price:
          type: number
          format: float
          example: 99.99
        category:
          type: string
          example: "Electronics"
        description:
          type: string
          example: "Description du produit"
        images:
          type: array
          items:
            type: string

    HealthCheck:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
          example: "healthy"
        services:
          type: object
          properties:
            legacySystem:
              type: string
              enum: [up, down, degraded]
              example: "up"
            produitService:
              type: string
              enum: [up, down, degraded]
              example: "up"
            venteService:
              type: string
              enum: [up, down, degraded]
              example: "up"
            stockService:
              type: string
              enum: [up, down, degraded]
              example: "up"
            reportingService:
              type: string
              enum: [up, down, degraded]
              example: "up"
        timestamp:
          type: string
          format: date-time

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key pour les intégrations externes
    
    ClientType:
      type: apiKey
      in: header
      name: X-Client-Type
      description: Type de client (pos, maisonmere, api, web)

security:
  - ApiKeyAuth: []
  - ClientType: []